Use deploymentform bindings and errors as react components
