--- dashboard/src/components/SelectRepoForm/index.tsx
@@ -1,3 +1,0 @@
-import SelectRepoForm from "./SelectRepoForm";
-
-export default SelectRepoForm;
